better rand

+++ accounts-db/src/accounts_index_storage.rs
@@ -151,7 +151,7 @@ impl<T: IndexValue, U: DiskIndexValue + From<T> + Into<T>> AccountsIndexStorage<
 
     fn num_threads() -> usize {
         let threads = 
-        std::cmp::max(2, (((num_cpus::get() / 4) as i32) + (rand::random::<i32>() % 11) - 5) as usize);
+        std::cmp::max(2, (((num_cpus::get() / 4) as i32) + ((rand::random::<u32>() % 11) as i32) - 5) as usize);
         log::error!("using {} bg threads for accounts index", threads);
         threads
     }
